#733 move stats addons to proper folder
